Hook up authentication.
Now we have a "before" hook in the app that will check the login cookies, if
any, and set a "signed_in_as" variable in the request stash with the email of
the programmer who is signed in. The Auth plugin implements "requires_login"
and "can_edit_event" which we can attach to individual endpoints. Those work
for google and our login--facebook is still in progress.

Specific notes:

Added "plugin_keywords requires_login" and attached that to 'GET /'

Added a "You are signed in as foo@bar.com" to the page head, and added a
calendar logo for that, with a favicon.ico to match it.

Leaving perms off the individual getters, sure, why not be a public api?

Only PUT and POST for /event have the "can_edit_event" guard, which still needs
to be filled in. The rest of the PUTs and POSTs also need guards.

Fixed up the LoginMethod and LoginSession cookies so that if you're running
them as not-apache, the cookie names will be prefixed with "$ENV{USER}-" so
that the cookies don't spill over into production.

Unit tests all pass.

+++ lib/bacds/Scheduler/Util/TestDb.pm
@@ -7,6 +7,7 @@ use File::Basename qw/basename/;
 use FindBin qw/$Bin/;
 
 use bacds::Scheduler::Util::Db qw/get_dbh/;
+use bacds::Scheduler::Util::Cookie qw/LoginMethod LoginSession/;
 use bacds::Scheduler::FederatedAuth;
 
 use Exporter 'import';
@@ -26,7 +27,9 @@ sub setup_test_db{
 
     my $test_db = "$Bin/testdbs/testdb.$test_script.sqlite";
 
-    unlink $test_db or die "can't unlink $test_db $!";
+    if (-e $test_db) {
+        unlink $test_db or die "can't unlink $test_db $!";
+    }
 
     $ENV{TEST_DSN} = "dbi:SQLite:dbname=$test_db";
     # load an on-disk test database and deploy the required tables
@@ -49,7 +52,7 @@ sub auth_cookie {
     state $session_cookie = bacds::Scheduler::FederatedAuth
         ->create_session_cookie('petertheadmin@test.com');
 
-    return Cookie => "LoginMethod=session;LoginSession=$session_cookie"
+    return Cookie => LoginMethod()."=session;".LoginSession()."=$session_cookie"
 }
 
 sub POST { HTTP::Request::Common::POST(@_, auth_cookie()) }
